requirements.txt
fastapi
torch
numpy
uvicorn
pillow
torchvision
lime
google-generativeai
